moving creating init.go in repositories
