+++ repositories/album_repositories.go
@@ -1,26 +1,12 @@
 package repositories
 
 import (
-	"example/web-service-gin/config"
 	"example/web-service-gin/models"
 	"fmt"
 	"net/http"
 
 	_ "github.com/go-sql-driver/mysql"
-	"github.com/jmoiron/sqlx"
 )
-
-var db *sqlx.DB
-var err error
-
-func init() {
-	if db == nil {
-		db, err = config.Connect()
-		if err != nil {
-			fmt.Println(err.Error())
-		}
-	}
-}
 
 type AlbumRepo struct {
 }
