+++ repositories/init.go
@@ -1,0 +1,20 @@
+package repositories
+
+import (
+	"example/web-service-gin/config"
+	"fmt"
+
+	"github.com/jmoiron/sqlx"
+)
+
+var db *sqlx.DB
+var err error
+
+func init() {
+	if db == nil {
+		db, err = config.Connect()
+		if err != nil {
+			fmt.Println(err.Error())
+		}
+	}
+}
